fixed PR comments from Yaron
