+++ PerimeterXModule/DataContracts/Requests/RiskRequest.cs
@@ -3,7 +3,6 @@
 namespace PerimeterX
 {
 	[DataContract]
-	[System.Obsolete("Use RiskRequestV2")]
 	public class RiskRequest
 	{
 		[DataMember(Name = "request")]
@@ -11,6 +10,9 @@ namespace PerimeterX
 
 		[DataMember(Name = "vid", EmitDefaultValue = false)]
 		public string Vid;
+
+		[DataMember(Name = "uuid", EmitDefaultValue = false)]
+		public string UUID;
 
 		[DataMember(Name = "additional", EmitDefaultValue = false)]
 		public Additional Additional;
