--- PerimeterXModule/DataContracts/Responses/RiskResponseScores.cs
@@ -1,17 +1,0 @@
-﻿using System.Runtime.Serialization;
-
-namespace PerimeterX
-{
-	[DataContract]
-	public class RiskResponseScores
-	{
-		[DataMember(Name = "filter")]
-		public int Filter;
-
-		[DataMember(Name = "suspected_script")]
-		public int Application;
-
-		[DataMember(Name = "non_human")]
-		public int Bot;
-	}
-}
