+++ PerimeterXModule/PerimeterXModule.csproj
@@ -96,16 +96,13 @@
     <Compile Include="PxModuleConfigurationSection.cs" />
     <Compile Include="Internals\ActivityReporter.cs" />
     <Compile Include="DataContracts\Requests\RiskRequestHeader.cs" />
-    <Compile Include="DataContracts\Requests\Request.cs" />
     <Compile Include="DataContracts\Requests\Additional.cs" />
-    <Compile Include="DataContracts\Requests\RiskRequest.cs" />
-    <Compile Include="DataContracts\Responses\RiskResponseScores.cs" />
     <Compile Include="DataContracts\Requests\CaptchaRequest.cs" />
     <Compile Include="DataContracts\Responses\CaptchaResponse.cs" />
-    <Compile Include="DataContracts\Requests\RiskRequestV2.cs" />
+    <Compile Include="DataContracts\Requests\RiskRequest.cs" />
     <Compile Include="DataContracts\Requests\RequestV2.cs" />
     <Compile Include="Internals\Validators\PXS2SValidator.cs" />
-    <Compile Include="DataContracts\Responses\RiskResponseV2.cs" />
+    <Compile Include="DataContracts\Responses\RiskResponse.cs" />
     <Compile Include="DataContracts\Activities\Activity.cs" />
     <Compile Include="DataContracts\Activities\ActivityDetails.cs" />
   </ItemGroup>
